BUILD : 1.7.6 (517) BUGS : 262 NOTES : Fixed integer overflow error with CS LEVELS.
git-svn-id: svn://svn.anope.org/anope/trunk@517 31f1291d-b8d6-0310-a050-a5561fc1590b


git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@371 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/chanserv.c
@@ -4882,7 +4882,7 @@ static int do_levels(User * u)
     char *error;
 
     ChannelInfo *ci;
-    short level;
+    int level;
     int i;
 
     /* If SET, we want two extra parameters; if DIS[ABLE] or FOUNDER, we want only
